requirements.txt
functions-framework==3.4.0
requests
google-cloud-secretmanager